chore(1-3389): move env accordion body
